requirements.txt
Flask==0.12
PyYAML==3.12
beautifulsoup4==4.5.1
lxml==3.6.4
psycopg2==2.6.2
requests==2.13.0
